Tests/integration/mqtt_test/requirements.txt
jsonschema
paho-mqtt
parameterized
docker